reporter/requirements.txt
# requirements.txt
google-cloud-firestore
google-cloud-pubsub
protobuf
